Develop d

+++ src/main/java/com/base/basesetup/dto/TdsCostInvoiceDTO.java
@@ -1,5 +1,7 @@
 package com.base.basesetup.dto;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -12,8 +14,8 @@ import lombok.NoArgsConstructor;
 public class TdsCostInvoiceDTO {
 	private Long id;
 	private String tdsWithHolding;
-	private float tdsWithHoldingPer;
+	private BigDecimal tdsWithHoldingPer;
 	private String section;
-	private int totTdsWhAmnt;
+	private BigDecimal totTdsWhAmnt;
 
 }
